Basic fish config

+++ setup.sh
@@ -15,10 +15,15 @@ for f in $OLDPWD/*; do
     if [[ "$O" == ".gitmodules" ]]; then continue; fi
     if [[ "$O" == "setup.sh" ]]; then continue; fi
     if [[ "$O" == "README.md" ]]; then continue; fi
+    if [[ "$O" == ".config" ]]; then continue; fi
 
     rm "$O" >/dev/null 2>&1
     ln -s "$f" "$O" >/dev/null 2>&1
 done
+
+mkdir ~/.config
+ln -s ~/.homedir/.config/fish ~/.config/fish
+
 ln -s .bashrc .bash_profile >/dev/null 2>&1
 ln -s .bashrc .bash_rc >/dev/null 2>&1
 ln -s .bashrc .profile >/dev/null 2>&1
